reorg customize functions
